+++ library_system/src/items/DVD.java
@@ -24,7 +24,7 @@ public class DVD implements LibraryItem{
     }
 
     @Override
-    public int calculateLateFee(int days) {
+    public double calculateLateFee(int days) {
         return days * 25;
     }
